split v0 and v1
